backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
python-multipart>=0.0.18
httpx>=0.28
pypdfium2>=4.0
Pillow>=11.0
pandas>=2.2
openpyxl>=3.1

python-docx>=1.1.0
reportlab>=4.0.0
pypdf>=3.17.0
img2pdf>=0.5.0
qrcode[pil]>=7.4.2
xhtml2pdf>=0.2.14
opencv-python-headless>=4.8.0
docx2pdf>=0.1.8
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
